ci: make test, clippy, and example builds pass
test job:
- Scope to --lib --tests --bins so CI doesn't require every example's
  runtime deps to be present
- Exclude tang-gpu (tests require a physical GPU adapter)
- Add tang-train as a dev-dep of tang-safetensors for its lib tests

clippy job:
- cargo clippy --fix applied the mechanical lint suggestions (manual
  assign, div_ceil, is_multiple_of, useless vec, etc.) across the
  workspace
- Hand-fix the remaining non-auto-fixable lints (unnecessary parens,
  clamp pattern, manual memcpy, same-item push, stripping prefix,
  approx_constant in production code)
- Add targeted #[allow(...)] on a few intentional patterns: NaN-
  preserving `!(x < tol)` comparisons, inherent `Dir3::neg`, Arc<Tape>
  with interior RefCell, approx_constant in test-data literals
- Gate four noisy stylistic lints in the CI invocation: too_many_args,
  type_complexity, needless_range_loop, ptr_arg — none of which fit
  numerical code conventions

tang-sheffer:
- Declare required-features = ["gpu"] on the gpu_search example so it
  is skipped when the feature is off
- Move `Atom` import into its test module (it was unused in lib code
  but needed by the test helpers)

+++ crates/tang-gpu/src/tensor.rs
@@ -107,7 +107,7 @@ impl GpuTensor {
         assert_eq!(self.ndim(), 2, "transpose requires 2D tensor");
         let m = self.shape[0] as u32;
         let n = self.shape[1] as u32;
-        let numel = (m * n) as u32;
+        let numel = m * n;
         let out = GpuTensor::uninit(device, &[n as usize, m as usize]);
 
         // Pack [count, M, N, 0] into params — reuse the existing params layout
